fix: Address is showing null values in customer details page
