--- app/src/main/java/com/mifos/apache/fineract/ui/online/customer/customerprofile/editcustomerprofilebottomsheet/EditAction.java
@@ -1,14 +1,0 @@
-package com.mifos.apache.fineract.ui.online.customer.customerprofile.editcustomerprofilebottomsheet;
-
-/**
- * @author Rajan Maurya
- *         On 07/08/17.
- */
-public enum  EditAction {
-
-    CAMERA,
-
-    GALLERY,
-
-    DELETE
-}
