+++ app/src/main/java/com/mifos/apache/fineract/ui/online/customers/customerprofile/editcustomerprofilebottomsheet/EditAction.java
@@ -1,0 +1,14 @@
+package com.mifos.apache.fineract.ui.online.customers.customerprofile.editcustomerprofilebottomsheet;
+
+/**
+ * @author Rajan Maurya
+ *         On 07/08/17.
+ */
+public enum  EditAction {
+
+    CAMERA,
+
+    GALLERY,
+
+    DELETE
+}
